Rename module `content-store-local` -> `file-store-local`.
It's not just about content anymore.
