+++ local-modules/hooks-server/Hooks.js
@@ -2,7 +2,7 @@
 // Licensed AS IS and WITHOUT WARRANTY under the Apache License,
 // Version 2.0. Details: <http://www.apache.org/licenses/LICENSE-2.0>
 
-import { LocalContentStore } from 'content-store-local';
+import { LocalContentStore } from 'file-store-local';
 import { Hooks as HooksCommon } from 'hooks-common';
 import { Singleton } from 'util-common';
